requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-Limiter>=3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0

# File processing
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
python-docx>=0.8.11
docx2txt>=0.8
reportlab>=4.0.0

# Offline AI dependencies
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
opencv-python>=4.8.0
numpy>=1.24.0

# Optional: Speech-to-Text (uncomment if needed)
# openai-whisper>=20231117
# SpeechRecognition>=3.10.0

# Fallback API dependencies (optional)
openai>=1.0.0
requests>=2.31.0
